Allow multiple text reveals to play at once and re-implement ColorRandomWriter.cs
Every 'text reveal' will subscribe themselves to the OnTick event located on the Textension.
The Textension script will invoke the event and allow each reveal to attempt to reveal the text.
String functionality has been removed from the individual text reveals and will have to be done to the textension class as a result.

Added some comments to different functions and variables

Deprecated the RenderTypeWriter.cs - This has been temporarily disabled