Add support to not hide the text on textension initialization
This will allow us to control whether or not we want to hide the text on initialization. If we do hide on initialization then we will probably reveal it using a TextReveal or other means. (I can see this being used with texts that have effects but don't need a specific reveal)