Created the beginnings of a Textensions Wizard / Tutorial component
This is an editor script that will target a TextensionWizard.cs (Which is just a Textension.cs with an enum to track state data), so we can show users step by step what needs to be hooked up and why. We can also show any warning if we can find a mistake using the previous statusIndicator asset. (I do want to expand on a status bar soon)